DEV: Update plugin to match latest guidelines
- Define extension modules
- Use different files instead of plugin.rb
- Make sure plugin is disabled according to the setting
